Remove build warnings in cryptography libraries
System.Security.Cryptography.Primitives uses System.IO-4.0.10.  The rest of the cryptography libraries seems to get an implicit reference to System.IO, but 4.0.0.  A post-build analysis is not pleased by this state of affairs.

So, for all libraries that currently package-to-package reference System.Security.Cryptography.Primitives; add a System.IO-4.0.10 reference.