Remove lazy initialization in raw errno -> Error conversion
It turns out that every place we get an ErrorInfo from a raw
errno, we access the PAL Error almost immediately. The lazy
initialization carried a risk that copies of an ErrorInfo
with the conversion deferred would get passed around and cause
the conversion to happen more than once. Also, there was an
unnecessary branch when accessing the Error.

The lazy initialization in the reverse Error -> raw errno case
is left intact because it is common to create an ErrorInfo from
an Error to pass to a caller that needs an ErrorInfo, but doesn't
always need the raw errno.