Short-circuiting versions of Enumerable methods.
Fixes #2349

Last<TSource>(this IEnumerable<TSource>, Func<TSource, bool>)
LastOrDefault<TSource>(this IEnumerable<TSource>, Func<TSource, bool>)
Single<TSource>(this IEnumerable<TSource>, Func<TSource, bool>)
SingleOrDefault<TSource>(this IEnumerable<TSource>, Func<TSource, bool>)
Min(this IEnumerable<float>)
Min(this IEnumerable<float?>)
Min(this IEnumerable<double>)
Min(this IEnumerable<double?>)

Last and LastOrDefault only short-circuit if the source is an IList<T>.

Last and LastOrDefault add a check for if the source is an IList<T> comparable
to that take by the form that don't take a predicate.

Apart from that no tests are added: While we could check for e.g. int.MaxValue
being seen by Max(this IEnumerable<int>), and so on, that would add a test for
every element, and so penalise sequences that did not contain it.

Min/MinOrDefault on double and float already had a similar improvement made in
5eb063ac, but with a backwards-compatibility loop added. Here that loop is
simply removed.

Included are tests for both the state before and after the changes.