Fix and enable a disabled Task test
The test was assuming that if all tasks queued to a scheduler have completed that the scheduler won't be making further use of whatever underlying scheduler it's targeting.  That assumption is false, however, as the scheduler may still queue additional housekeeping tasks, and it requires that the underlying scheduler remain alive until the overlaying scheduler declares completion via its Completion task.