Harden X509Certificate collections
This commit introduces minor behavior changes in order to harden
the X509Certificate collections in .NET Core:

1. Using X509CertificateCollection or X509Certificate2Collection as
an IList allows adding/inserting non-X509Certificate objects into the
collection. Subsequent calls to enumerate or access these items
results in InvalidCastException being thrown. This matches the full
.NET Framework (desktop) behavior. Clearly this is not the intended
behavior of these collections, which are meant to hold
X509Certificate and X509Certificate2 objects. This commit changes
X509CertificateCollection and X509Certificate2Collection to throw
ArgumentException when non-X509Certificate objects are attempted to
be added/inserted, via List<X509Certificate>. As part of this,
cleanup calling the non-generic IList implementations and null
checking.

2. X509CertificateEnumerator has a public constructor with a
collection parameter that it uses without first verifying the
parameter is non-null, which could result in NullReferenceExceptions.
This commit adds a null check, throwing ArgumentNullException if the
parameter is null.