X509Certificate2Collection.Contains: Allow null certificates
X509Certificate2Collection.Contains throws ArgumentNullException when
a null value is passed in, but the IList.Contains implementation and
X509CertificateCollection.Contains do not. This commit changes
X509Certificate2Collection.Contains to not throw for a null value to
be consistent with the IList.Contains implementation and
X509CertificateCollection.Contains.