Avoid test deadlocks in threading WaitHandle tests
A bunch of the Mutex/EventWaitHandle/Semaphore tests are waiting indefinitely for a signal.  If that signal never arrives due to the test failing, CI will end up waiting for several hours and then aborting the whole run.  This change adds timeouts to those waits.  It also changes a name used in one of the semaphore tests to be randomly generated; this should hopefully fix a pervasive failure we're seeing in CI right now.