Make registry tests more robust
Currently if the registry tests are interrupted during their execution, they leave behind state in the registry that causes all subsequent runs of the tests to fail. This commit cleans up the registry tests to avoid this.  The logic for creating a name for the test class is moved into the base, which also deletes any existing key of that name prior to running a test.