Restore virtual generic usage in LINQ
During the bring-up of .NET Native, there was a period of time during which virtual generic methods were not supported.  As a result, LINQ's implementation was changed to avoid the usage of virtual generics, which occur in some optimizations around Select usage.  That limitation no longer exists.  This commit restores the original code.