Fix order-dependent TypeExtensions tests
These tests were expected to be run in sequential order, as they all
modified a shared static variable on some class. Instead, they now each
modify a separate variable and should therefore be order-independent.
There was no need for them to be sharing that state; it wasn't important
to the test.