Refactor PipeStream on Unix for isAsync
Right now the PipeStream implementation on Unix implements isAsync by delegating to the base Stream class, the same as it does for !isAsync.  This limits our ability to do anything "special" in the async implementation, such as supporting cancellation more fully.  This commit refactors things to avoid depending on the base Stream when isAsync is true (we still depend on it when isAsync is false, as we do on Windows).

In doing so, we clean up the Windows implementation as well, moving more shared logic into PipeStream.cs and out of .Windows.cs/.Unix.cs.  This helps make the *Core methods very focused on their specific task.  It also fixes an issue that existed in the Unix implementation but was solved in the Windows implementation.  On Windows, we ensure that if you create the stream as synchronous, you always go through the sync methods: if you call the Read/Write methods, they do the writing synchronously, and if you call the ReadAsync/WriteAsync methods, they use the base Stream class to queue work items that use Read/Write.  If you create the stream as asynchronous, you always go through the async methods: if you call ReadAsync/WriteAsync, you get the specialized async functionality, and if you call Read/Write, they delegate to the ReadAsync/WriteAsync functions and block waiting for them to complete.  On Unix, we weren't doing this, which leads to a problem if you mix/match Read/Write with ReadAsync/WriteAsync methods: on Windows on an async stream, that works fine, but on Unix you'd end up with potentially concurrent calls to Read/Write, which is not safe and can end up corrupting the instance.  This change modifies the Unix implementation to use the same scheme as on Windows, such that you can't end up in this situation on an async stream.