Separate ToArray specialisation from Enumerable.Iterator
Enumerable.Iterator has a virtual ToArray that is overridden to provide
optimised array-creation & buffer-filling when possible.

Most Enumerable.Iterator subclasses don't override the default, which
results in it passing through Buffer's constructor a second time. This can
also happen if those that do override it have a predicate.

Separate out this part of Iterator's interface into a separate interface, for
only those classes that override it, and signal a failure to create an array
by returning null.