Fix multiple MSB3277 build warnings
Fix several warnings of the form:
```
warning MSB3277: Found conflicts between different versions of the s
ame dependent assembly that could not be resolved.  These reference conflicts are listed in the build log when log verbosity is set to detailed.
```
Brings the number of warnings during a full build back down to 0.