Avoid blocking CurlHandler write callback
When receiving the body of an HTTP response, libcurl provides the data
received via a callback that's invoked synchronously from the thread
pumping libcurl's processing.  This callback can't return until it's
handed off all of the available data, and thus the current implementation
of CurlHandler blocks that thread until a reader has consumed the data.
This is both expensive and dangerous.  It's expensive because we end up
blocking a thread until some reader comes along, and it's dangerous
because this can easily lead to deadlock.  A libcurl multi handle may be
used by only one thread at a time, which means that if the handle is being
used by a thread while blocked on the write callback, no other thread can
access the multi handle, which means no other callbacks for other
concurrent requests can be invoked.  If multiple operations were queued,
and then the data for them arrived in an order that didn't exactly match
the order of readers consuming the data, deadlock would ensue.

This change makes it so that we no longer block the callback, taking
advantage of a pausing mechanism libcurl provides that allows one to pause
a connection until it's needed.

I rewrote the CurlResponseStream to manage this.  Now when the callback
has data to write, it attempts to transfer it to the stream.  If there's a
reader already waiting for the data (represented by a task stored on the
stream), the callback simply transfers its data to the reader.  If the
callback has too much data than would fit in the reader, the remaining
data is buffered on the stream.  If instead the callback finds that either
there isn't a waiting task or that there's already some remaining data
buffered from a previous callback, it instead pauses the connection, such
that no more callbacks will be invoked until we unpause.  From the reader
side, a reader who tries to read data from the stream can pull data from
the remaining buffer if there is any, or can register a task with the
stream that will be filled with data at some point when callback has data
available.

As part of debugging some issues with this change, I added a verbose
tracing mechanism that's enabled by compiling the library with the
CURLHANDLER_VERBOSE constant (the same that we use for enabling libcurl's
verbose diagnostic output).  I left the tracing in as I found it valuable
in diagnosing issues with the library, but we can remove it later if
desired.

I also added a test that deadlocks before the change and that completes
successfully after.