Comment out ctrl-C System.Console test
Our test for ctrl-C handling on Unix generally works fine, but it causes problems when run in xunit, which has its own ctrl-C handler that stops running tests. Since the test doesn't make sense for now but we might want something like it later, I'm commenting it out (added a comment to the code as to why it's commented out) rather than marking it as [ActiveIssue] or deleting it.