Redirect handling for cookies and creds
This fix realigns CurlHandler with WinHttpHandler w.r.t Redirect handling.
When redirect is detected, we clear the credentials if the credentials were originally provided as NetworkCredentials Otherwise we use the
forwarding URI to get the new set of credentials.
Like wise, in case of redirects, we get the cookies again from the cookieContainer

In addition, this commit also collects the cookies from the response header