Remove U+037F from "Unassigned" oracle in test.
The Char casing test has its own copy of some Unicode data which has a
list of unasigned codepoints.  The goal of the test is to ensure that
these items case to themselves.  However, U+037F is now assigned as of
Unicode 7.0 (it is GREEK CAPITAL LETTER YOT) so a system which does
casing according to the Unicode 7 standard will actually fail this test,
since it will correctly case this letter to U+03F3.