Fix incorrect codegen for self-application of recursive expressions.
ET binder associates scope information with corresponding lambda nodes, which is valid in most cases since variables that belong to lambda depend only on the lambda contents.
Except that when lambda is inlined and parameters become locals.

This change makes scopes for inlined lambdas to be associated with invocation nodes, which disambiguates them from scopes for lambdas in their regular non-inlined form.

Fixes: #3224