Add platform selection to build.sh.
Prior to recent changes, it was possible to use build.sh to cross-compile
CoreFX by passing "/p:OSGroup=<target OS>" at the command line. This
change restores the ability to cross-compile by allowing the specification
of the target OS as a command-line option to build.sh: any one of
"windows", "linux", "osx", or "freebsd" is accepted.