Unix X509Chains treats root certs in ExtraStore differently than Windows.
When building an X509Chain on Unix, if the root cert isn't in the system store, the building still succeeds with no errors for UntrustedRoot.  This is different than how it works on Windows.

I changed OpenSslX509ChainProcessor.BuildChain to check if the root cert is contained in the LocalMachine root or CurrentUser root store.  If it isn't, add an UntrustedRoot error.

Fix #3226.