Implement IDNA2008 on top of ICU
This change uses the ToUnicode and ToAscii methods from
System.Globalization.Native which are implemented on top of ICU's UTS46
implementation.

There are some test issues that need to be addressed due to differences
between Windows implementation of the standard and ICUs (which follows
the standard more closely). These will be addressed at a later date. I
have confirmed manually that things are working as expected.

Fixes #2779
Fixes #810