OpenSSL multi-threading support on OSX
In order to support multi-threading in OpenSSL on OSX, we need to set the CRYPTO_set_id_callback to a method that will return the current thread id.

This is necessary because OSX uses a version of OpenSSL less than 1.0.  Before 1.0, setting this callback was necessary because the default uses getpid(), which will be the same value for multiple threads in the same process.  Any version 1.0 or greater of OpenSSL will use the address of errno to differentiate between threads.

Fix #2210