Fix seg faults due to incorrect crypto DllImport signatures
Three DllImports were declared to return strings rather than IntPtr.  Returning a string like this should only be done when the native code is allocating a ```char*``` that must be freed by the consumer, and the CLR dutifully attempts to do so after marshaling the ```char*``` to the managed string.  However, these three DllImports all return ```char*```s that shouldn't be freed, which was resulting in seg faults.

The fix is simply to declare the DllImports to return IntPtr, and then use Marshal.PtrToStringAnsi to do the marshaling.