Delete two PLINQ tests
Enumerators, even those produced by PLINQ, are not safe to be used concurrently from multiple threads.  Two PLINQ tests are doing exactly that, trying to Dispose of an enumerator concurrently with calling MoveNext on it and potentially concurrently with other calls to Dispose.  This is resulting in sporadic hangs in CI.

I'm simply deleting the tests, as their sole purpose is to test this particular unsupported and unsafe situation.