Use custom test-runtime project.json
Today, the test-runtime project.json and project.lock.json we have comes
from the buildtools package. This is less than ideal for a few reasons:

 - To change anything in here requires that we roll a new version of
   build tools and consume that.
 - The version in buildtools does not have runtime sections for either
   Ubuntu 14.04 or OSX 10.10, so when we restore a test runtime we
   always get the windows version. This leads us to having to do upstack
   hacks like "run-test.sh" which tries to replace the windows versions
   of things with correct cross platform versions.

Now that we are publishing cross platform packages, we should be using
them when we run tests. This change introduces a copy of the test
runtime project.json we can version at a different cadence than build
tools and updates the build to use it instead.

One artifact of this change is you can now pass
/p:TestNugetRuntimeId="ubuntu.14.04-x64" to the build and the test
folders produced will actually be runable on Ubuntu 14.04 (and you could
do a similar thing for osx.10.10-x64).

The next step will be to start ripping out code in run-test.sh so we
don't actually do a bunch of extra copies we don't need (e.g. we can
remove all of the CoreFX managed copying).

This also moves us one step closer to being able to easily update the
version of the runtime we test against.