Add missing runtimes to project.json
System.Diagnostics.TextWriterTraceListener.Tests has some dependencies
on System.Diagnositcs.TraceSource which has a different version per
runtime (due to the fact that it can't run in UWP, I believe). We need
to list runtimes in our project.json files so the correct platform
speficic assets are deployed when we build out the test drop.