Don't override dir.props' ECMA Key setting
Due to https://github.com/dotnet/roslyn/issues/2444, we have disabled
ECMA signing when the host platform for our build is not Windows. This
is done by explicitly setting the property in our top level dir.props.
Some of the reference assemblies explicitly set UseECMAKey to true,
overwriting this value which leads to build breaks when building on non
Windows platforms.

Guard the setting of this property so that we don't override it if it
already set.

Fixes #3739